requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
dj-email-url==1.0.1
django==3.0.8
django-appconf==1.0.4
djangorestframework==3.12.2
django-cache-url==3.1.2
environs==8.0.0
gunicorn==20.0.4
marshmallow==3.7.0
psycopg2-binary==2.8.5
python-dotenv==0.14.0
pytz==2020.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.3.1
unipath==1.1
whitenoise==5.1.0
pytest==5.4.3
pytest-django==3.10.0
pytest-mock==3.5.1
django-webpack-loader==0.7.0
django_extensions==3.1.1
